chore: added custom hook for obtaining game catalog
