+++ src/components/GameGallery/GameGallery.tsx
@@ -1,31 +1,8 @@
-import { useEffect, useState } from "react";
-
-import apiClient from "../../services/api-client";
-
-interface Game {
-    id: number,
-    name: string
-}
-
-interface APIResponse {
-    amount: number,
-    results: Game[]
-}
+import useGameCatalog from "@hooks/useGameCatalog";
 
 const GameGallery = () => {
 
-    const [ games, setGames ] = useState<Game[]>([]);
-
-    const [ error, setError ] = useState('');
-
-    console.log(games)
-
-    useEffect(() => {
-
-        apiClient.get<APIResponse>('/games')
-            .then(response => setGames(response.data.results))
-            .catch(error => setError(error.message))
-    }, [])
+    const { games, error } = useGameCatalog();
 
     return (
         <ul>
