+++ vite.config.ts
@@ -8,7 +8,9 @@ export default defineConfig({
         alias: {
             '@components': '/src/components',
             '@assets': '/src/assets',
-            '@common': '/src/common'
+            '@common': '/src/common',
+            '@services': '/src/services',
+            '@hooks': '/src/hooks'
         },
     },
 })
